locust_performance/requirements.txt
boto3==1.28.1
botocore==1.31.1
dpath==2.1.6
locustio==0.999
tinydb==4.8.0
